[#4661] Isolate domain model and transport model

+++ commons-server/src/test/java/com/navercorp/pinpoint/common/server/bo/SpanFactoryAssert.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2018 NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.navercorp.pinpoint.common.server.bo;
 
 import com.navercorp.pinpoint.common.util.TransactionId;
@@ -141,15 +157,15 @@ public class SpanFactoryAssert {
     private void assertSpanEventList(List<SpanEventBo> spanEventBoList, List<TSpanEvent> spanEventList) {
         Assert.assertEquals(CollectionUtils.isEmpty(spanEventBoList), CollectionUtils.isEmpty(spanEventList));
         if (CollectionUtils.isNotEmpty(spanEventBoList)) {
-            Map<Long, SpanEventBo> spanEventBoMap = new HashMap<Long, SpanEventBo>();
+            Map<Integer, SpanEventBo> spanEventBoMap = new HashMap<Integer, SpanEventBo>();
             for (int i = 0; i < spanEventBoList.size(); i++) {
                 SpanEventBo spanEventBo = spanEventBoList.get(i);
-                spanEventBoMap.put((long)spanEventBo.getSequence(), spanEventBo);
+                spanEventBoMap.put((int)spanEventBo.getSequence(), spanEventBo);
             }
 
             for (int i = 0; i < spanEventList.size(); i++) {
                 TSpanEvent tSpanEvent = spanEventList.get(i);
-                SpanEventBo spanEventBo = spanEventBoMap.get((long) tSpanEvent.getSequence());
+                SpanEventBo spanEventBo = spanEventBoMap.get((int) tSpanEvent.getSequence());
                 Assert.assertNotNull(spanEventBo);
                 assertSpanEvent(tSpanEvent, spanEventBo);
             }
